Refactor engine services into focused modules
